Moved hashing to utils and added wipe cache cmd
